Moving paparazzi tests to separate module.
